feat: much improved schema support
This is what I originally wanted to do but deemed it too much work and
too much of a breaking change at the time. Basically, schema information
is encoded as expressions like so:

    attr = string(optional enum("a", "b", "c") default("dflt"))

Repeated blocks are encoded similarly:

    block(repeated) { ... }

Also removed JSON support. It never quite worked correctly, and I think
we can just rely on Go's JSON support for now.